refactor: move private statics to internal file
