+++ packages/opentelemetry-plugin-grpc/package.json
@@ -43,7 +43,7 @@
     "@types/protobufjs": "^6.0.0",
     "@types/shimmer": "^1.0.1",
     "codecov": "^3.5.0",
-    "grpc": "1.22.2",
+    "grpc": "^1.23.3",
     "gts": "^1.1.0",
     "mocha": "^6.2.0",
     "nyc": "^14.1.1",
