+++ packages/opentelemetry-plugin-grpc/src/enums/AttributeNames.ts
@@ -15,7 +15,7 @@
  */
 
 export enum GrpcAttributeNames {
-  COMPONENT = 'grpc.component',
+  COMPONENT = 'component',
   GRPC_KIND = 'grpc.kind', // SERVER or CLIENT
   GRPC_METHOD = 'grpc.method',
   GRPC_STATUS_CODE = 'grpc.status_code',
